refactoring CLI add create database command
