+++ lib/build-folders/endpoint.js
@@ -31,7 +31,7 @@ module.exports = function( {
         "src/controller",
         "views/main",
         "public/assets",
-        "bin",
+        "bin/commands",
         "mvc-vendor",
         "config"
     ].forEach( ressource => {
